+++ sdk/src/protocols/aave/v3/borrow.test.ts
@@ -3,8 +3,8 @@ import { avatar, member } from "../../../../test/wallets"
 import { applyPermissions, stealErc20 } from "../../../../test/helpers"
 import { contracts } from "../../../../eth-sdk/config"
 import { Status } from "../../../../test/types"
-import { testKit } from "../../../../test/kit"
-import { parseEther, parseUnits } from "ethers/lib/utils"
+import kit from "../../../../test/kit"
+import { parseEther, parseUnits } from "ethers"
 
 describe("aave_v3", () => {
   describe("borrow", () => {
@@ -20,16 +20,16 @@ describe("aave_v3", () => {
         contracts.mainnet.balancer.vault
       )
       await expect(
-        testKit.eth.usdc.approve(
+        kit.asMember.usdc.approve(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseUnits("10000", 6)
         )
       ).not.toRevert()
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.supply(
+        kit.asMember.aaveV3.aaveLendingPoolV3.supply(
           contracts.mainnet.usdc,
           parseUnits("10000", 6),
-          avatar._address,
+          avatar.address,
           0
         )
       ).not.toRevert()
@@ -37,27 +37,25 @@ describe("aave_v3", () => {
 
     it("borrow ETH and repay", async () => {
       await expect(
-        testKit.eth.aaveV3.variableDebtWETH.approveDelegation(
+        kit.asMember.aaveV3.variableDebtWETH.approveDelegation(
           contracts.mainnet.aaveV3.wrappedTokenGatewayV3,
           parseEther("1")
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.aaveV3.wrappedTokenGatewayV3.borrowETH(
+        kit.asMember.aaveV3.wrappedTokenGatewayV3.borrowETH(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseEther("1"),
-          2,
           0
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.aaveV3.wrappedTokenGatewayV3.repayETH(
+        kit.asMember.aaveV3.wrappedTokenGatewayV3.repayETH(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseEther("0.5"),
-          2,
-          avatar._address,
+          avatar.address,
           { value: parseEther("0.5") }
         )
       ).not.toRevert()
@@ -65,37 +63,37 @@ describe("aave_v3", () => {
 
     it("deposit ETH, borrow USDC and repay", async () => {
       await expect(
-        testKit.eth.aaveV3.wrappedTokenGatewayV3.depositETH(
+        kit.asMember.aaveV3.wrappedTokenGatewayV3.depositETH(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
-          avatar._address,
+          avatar.address,
           0,
           { value: parseEther("1") }
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.borrow(
+        kit.asMember.aaveV3.aaveLendingPoolV3.borrow(
           contracts.mainnet.usdc,
           parseUnits("100", 6),
           2,
           0,
-          avatar._address
+          avatar.address
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.usdc.approve(
+        kit.asMember.usdc.approve(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseUnits("50", 6)
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.repay(
+        kit.asMember.aaveV3.aaveLendingPoolV3.repay(
           contracts.mainnet.usdc,
           parseUnits("50", 6),
           2,
-          avatar._address
+          avatar.address
         )
       ).not.toRevert()
     })
@@ -104,17 +102,16 @@ describe("aave_v3", () => {
     // Test with ETH
     it("allows borrowing ETH from avatar", async () => {
       await expect(
-        testKit.eth.aaveV3.variableDebtWETH.approveDelegation(
+        kit.asMember.aaveV3.variableDebtWETH.approveDelegation(
           contracts.mainnet.aaveV3.wrappedTokenGatewayV3,
           parseEther("1")
         )
       ).toBeAllowed()
 
       await expect(
-        testKit.eth.aaveV3.wrappedTokenGatewayV3.borrowETH(
+        kit.asMember.aaveV3.wrappedTokenGatewayV3.borrowETH(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseEther("1"),
-          2,
           0
         )
       ).toBeAllowed()
@@ -122,21 +119,19 @@ describe("aave_v3", () => {
 
     it("only allows repaying ETH from avatar", async () => {
       await expect(
-        testKit.eth.aaveV3.wrappedTokenGatewayV3.repayETH(
+        kit.asMember.aaveV3.wrappedTokenGatewayV3.repayETH(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseEther("1"),
-          2,
-          avatar._address,
+          avatar.address,
           { value: parseEther("1") }
         )
       ).toBeAllowed()
 
       await expect(
-        testKit.eth.aaveV3.wrappedTokenGatewayV3.repayETH(
+        kit.asMember.aaveV3.wrappedTokenGatewayV3.repayETH(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseEther("1"),
-          2,
-          member._address,
+          member.address,
           { value: parseEther("1") }
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
@@ -144,7 +139,7 @@ describe("aave_v3", () => {
 
     it("allows swapping the ETH borrow rate mode", async () => {
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.swapBorrowRateMode(
+        kit.asMember.aaveV3.aaveLendingPoolV3.swapBorrowRateMode(
           contracts.mainnet.weth,
           1
         )
@@ -154,22 +149,22 @@ describe("aave_v3", () => {
     // Test with USDC
     it("only allows borrowing USDC from avatar", async () => {
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.borrow(
+        kit.asMember.aaveV3.aaveLendingPoolV3.borrow(
           contracts.mainnet.usdc,
           parseUnits("10000", 6),
           2,
           0,
-          avatar._address
+          avatar.address
         )
       ).toBeAllowed()
 
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.borrow(
+        kit.asMember.aaveV3.aaveLendingPoolV3.borrow(
           contracts.mainnet.usdc,
           parseUnits("10000", 6),
           2,
           0,
-          member._address
+          member.address
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
     })
@@ -182,34 +177,34 @@ describe("aave_v3", () => {
       )
 
       await expect(
-        testKit.eth.usdc.approve(
+        kit.asMember.usdc.approve(
           contracts.mainnet.aaveV3.aaveLendingPoolV3,
           parseUnits("10000", 6)
         )
       ).toBeAllowed()
 
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.repay(
+        kit.asMember.aaveV3.aaveLendingPoolV3.repay(
           contracts.mainnet.usdc,
           parseUnits("10000", 6),
           2,
-          avatar._address
+          avatar.address
         )
       ).toBeAllowed()
 
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.repay(
+        kit.asMember.aaveV3.aaveLendingPoolV3.repay(
           contracts.mainnet.usdc,
           parseUnits("10000", 6),
           2,
-          member._address
+          member.address
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
     })
 
     it("allows swapping the USDC borrow rate mode", async () => {
       await expect(
-        testKit.eth.aaveV3.aaveLendingPoolV3.swapBorrowRateMode(
+        kit.asMember.aaveV3.aaveLendingPoolV3.swapBorrowRateMode(
           contracts.mainnet.usdc,
           1
         )