+++ sdk/src/protocols/rocket_pool/utils.ts
@@ -1,20 +1,23 @@
-import { getMainnetSdk } from "@dethcrypto/eth-sdk-client"
+import { getMainnetSdk } from "@gnosis-guild/eth-sdk-client"
 import { ethProvider } from "../../provider"
 import { getProvider } from "../../../test/provider"
-import { utils } from "ethers"
+import { keccak256, solidityPacked } from "ethers"
 
 const sdk = getMainnetSdk(
   process.env.NODE_ENV === "test" ? getProvider() : ethProvider
 )
 
 export const queryDepositPool = async () => {
-  const deposit_pool_key = utils.keccak256(
-    utils.solidityPack(
+  const deposit_pool_key = keccak256(
+    solidityPacked(
       ["string", "string"],
       ["contract.address", "rocketDepositPool"]
     )
   )
-  return (await sdk.rocket_pool.storage.getAddress(
+
+  // TODO: we need this as any cast because typechain does not yet correctly generate the types for conflicting function names
+  // (getAddress is a BaseContract member, so the full TypeChain would have to create the function member type under the full signature key)
+  return (await (sdk.rocket_pool.storage as any)["getAddress(bytes32)"](
     deposit_pool_key
   )) as `0x${string}`
 }