+++ sdk/src/protocols/spark/deposit.test.ts
@@ -3,15 +3,11 @@ import { avatar, member } from "../../../test/wallets"
 import { applyPermissions, stealErc20 } from "../../../test/helpers"
 import { contracts } from "../../../eth-sdk/config"
 import { Status } from "../../../test/types"
-import { testKit } from "../../../test/kit"
-import { getMainnetSdk } from "@dethcrypto/eth-sdk-client"
-import { parseEther, parseUnits } from "ethers/lib/utils"
-import { ethers } from "ethers"
+import kit from "../../../test/kit"
+import { formatUnits, parseEther, parseUnits } from "ethers"
 
 const MAX_AMOUNT =
   115792089237316195423570985008687907853269984665640564039457584007913129639935n
-
-const sdk = getMainnetSdk(avatar)
 
 describe("spark", () => {
   describe("deposit", () => {
@@ -24,18 +20,18 @@ describe("spark", () => {
     // Test with ETH
     it("only allows depositing ETH on behalf of avatar", async () => {
       await expect(
-        testKit.eth.spark.wrappedTokenGatewayV3.depositETH(
+        kit.asMember.spark.wrappedTokenGatewayV3.depositETH(
           contracts.mainnet.spark.sparkLendingPoolV3,
-          avatar._address,
+          avatar.address,
           0,
           { value: parseEther("1") }
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.wrappedTokenGatewayV3.depositETH(
+        kit.asMember.spark.wrappedTokenGatewayV3.depositETH(
           contracts.mainnet.spark.sparkLendingPoolV3,
-          member._address,
+          member.address,
           0,
           { value: parseEther("1") }
         )
@@ -44,45 +40,45 @@ describe("spark", () => {
 
     it("only allows withdrawing ETH from avatars' position", async () => {
       await expect(
-        testKit.eth.spark.spWETH.approve(
+        kit.asMember.spark.spWETH.approve(
           contracts.mainnet.spark.wrappedTokenGatewayV3,
           parseEther("1")
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.wrappedTokenGatewayV3.withdrawETH(
+        kit.asMember.spark.wrappedTokenGatewayV3.withdrawETH(
           contracts.mainnet.spark.sparkLendingPoolV3,
           parseEther("1"),
-          avatar._address
+          avatar.address
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.wrappedTokenGatewayV3.withdrawETH(
+        kit.asMember.spark.wrappedTokenGatewayV3.withdrawETH(
           contracts.mainnet.spark.sparkLendingPoolV3,
           parseEther("1"),
-          member._address
+          member.address
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
     })
 
     it("allow setting the deposited ETH as collateral", async () => {
       let reserve_config: Array<any> =
-        await sdk.spark.data_provider.getReserveConfigurationData(
+        await kit.asAvatar.spark.data_provider.getReserveConfigurationData(
           contracts.mainnet.weth
         )
       const collateralizable: boolean = reserve_config[5]
       if (collateralizable) {
         await expect(
-          testKit.eth.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
+          kit.asMember.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
             contracts.mainnet.weth,
             true
           )
         ).not.toRevert()
       } else {
         await expect(
-          testKit.eth.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
+          kit.asMember.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
             contracts.mainnet.weth,
             true
           )
@@ -98,26 +94,26 @@ describe("spark", () => {
         contracts.mainnet.balancer.vault
       )
       await expect(
-        testKit.eth.weth.approve(
+        kit.asMember.weth.approve(
           contracts.mainnet.spark.sparkLendingPoolV3,
           parseEther("1")
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.supply(
+        kit.asMember.spark.sparkLendingPoolV3.supply(
           contracts.mainnet.weth,
           parseEther("1"),
-          avatar._address,
+          avatar.address,
           0
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.supply(
+        kit.asMember.spark.sparkLendingPoolV3.supply(
           contracts.mainnet.weth,
           parseEther("1"),
-          member._address,
+          member.address,
           0
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
@@ -125,18 +121,18 @@ describe("spark", () => {
 
     it("only allows withdrawing WETH from avatars' position", async () => {
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.withdraw(
+        kit.asMember.spark.sparkLendingPoolV3.withdraw(
           contracts.mainnet.weth,
           parseEther("1"),
-          avatar._address
+          avatar.address
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.withdraw(
+        kit.asMember.spark.sparkLendingPoolV3.withdraw(
           contracts.mainnet.weth,
           parseEther("1"),
-          member._address
+          member.address
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
     })
@@ -144,19 +140,19 @@ describe("spark", () => {
     // Claim rewards to avatar
     it("only claim rewards to avatar", async () => {
       await expect(
-        testKit.eth.spark.RewardsController.claimRewards(
+        kit.asMember.spark.RewardsController.claimRewards(
           [contracts.mainnet.spark.spWETH],
           MAX_AMOUNT,
-          avatar._address,
+          avatar.address,
           contracts.mainnet.lido.wsteth
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.RewardsController.claimRewards(
+        kit.asMember.spark.RewardsController.claimRewards(
           [contracts.mainnet.spark.spWETH],
           MAX_AMOUNT,
-          member._address,
+          member.address,
           contracts.mainnet.lido.wsteth
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
@@ -170,26 +166,26 @@ describe("spark", () => {
         contracts.mainnet.balancer.vault
       )
       await expect(
-        testKit.eth.usdc.approve(
+        kit.asMember.usdc.approve(
           contracts.mainnet.spark.sparkLendingPoolV3,
           parseUnits("1000", 6)
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.supply(
+        kit.asMember.spark.sparkLendingPoolV3.supply(
           contracts.mainnet.usdc,
           parseUnits("1000", 6),
-          avatar._address,
+          avatar.address,
           0
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.supply(
+        kit.asMember.spark.sparkLendingPoolV3.supply(
           contracts.mainnet.usdc,
           parseUnits("1000", 6),
-          member._address,
+          member.address,
           0
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
@@ -197,38 +193,38 @@ describe("spark", () => {
 
     it("only allows withdrawing USDC from avatars' position", async () => {
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.withdraw(
+        kit.asMember.spark.sparkLendingPoolV3.withdraw(
           contracts.mainnet.usdc,
           parseUnits("1000", 6),
-          avatar._address
+          avatar.address
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sparkLendingPoolV3.withdraw(
+        kit.asMember.spark.sparkLendingPoolV3.withdraw(
           contracts.mainnet.usdc,
           parseUnits("1000", 6),
-          member._address
+          member.address
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
     })
 
     it("allow setting the deposited USDC as collateral", async () => {
       let reserve_config: Array<any> =
-        await sdk.spark.data_provider.getReserveConfigurationData(
+        await kit.asAvatar.spark.data_provider.getReserveConfigurationData(
           contracts.mainnet.usdc
         )
       const collateralizable: boolean = reserve_config[5]
       if (collateralizable) {
         await expect(
-          testKit.eth.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
+          kit.asMember.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
             contracts.mainnet.usdc,
             true
           )
         ).not.toRevert()
       } else {
         await expect(
-          testKit.eth.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
+          kit.asMember.spark.sparkLendingPoolV3.setUserUseReserveAsCollateral(
             contracts.mainnet.usdc,
             true
           )
@@ -243,34 +239,34 @@ describe("spark", () => {
         contracts.mainnet.balancer.vault
       )
       await expect(
-        testKit.eth.dai.approve(
+        kit.asMember.dai.approve(
           contracts.mainnet.spark.sDAI,
           parseEther("1000")
         )
       ).not.toRevert()
       await expect(
-        testKit.eth.spark.sDAI.deposit(parseEther("1000"), avatar._address)
+        kit.asMember.spark.sDAI.deposit(parseEther("1000"), avatar.address)
       ).not.toRevert()
-      const sdai_balance_bn = await sdk.spark.sDAI.balanceOf(avatar._address)
-      const sdai_balance = ethers.utils
-        .formatUnits(sdai_balance_bn, 18)
-        .toString()
+      const sdai_balance_bn = await kit.asAvatar.spark.sDAI.balanceOf(
+        avatar.address
+      )
+      const sdai_balance = formatUnits(sdai_balance_bn, 18).toString()
       await expect(
-        testKit.eth.spark.sDAI.redeem(
+        kit.asMember.spark.sDAI.redeem(
           parseEther(sdai_balance),
-          avatar._address,
-          avatar._address
+          avatar.address,
+          avatar.address
         )
       ).not.toRevert()
 
       await expect(
-        testKit.eth.spark.sDAI.deposit(parseEther("1000"), member._address)
+        kit.asMember.spark.sDAI.deposit(parseEther("1000"), member.address)
       ).toBeForbidden(Status.ParameterNotAllowed)
       await expect(
-        testKit.eth.spark.sDAI.redeem(
+        kit.asMember.spark.sDAI.redeem(
           parseEther("1000"),
-          member._address,
-          member._address
+          member.address,
+          member.address
         )
       ).toBeForbidden(Status.ParameterNotAllowed)
     })